### 2025-03-07 #7884 @editorialbot
[PRE REVIEW]: Lighter: Configuration-Driven Deep Learning
**Submitting author:** @ibro45 (<a href="http://orcid.org/0000-0002-8397-5940">Ibrahim Hadzic</a>)
**Repository:** https://github.com/project-lighter/lighter
**Branch with paper.md** (empty if default branch):
**Version:** v0.0.3a20
