### 2025-03-08 082c551
Add ORCID ids and missing DOIs to the paper. Shorten and simplify the language in the paper.
