+++ paper/paper.md
@@ -8,20 +8,24 @@ tags:
   - framework
 authors:
   - name: Ibrahim Hadzic
-    orcid: 0000-0000-0000-0000
+    orcid: 0000-0002-8397-5940
     corresponding: true
     affiliation: "1, 2"
   - name: Suraj Pai
-    orcid: 0000-0000-0000-0000
+    orcid: 0000-0001-8043-2230
     affiliation: "2, 3, 4"
   - name: Keno Bressem
     affiliation: "5, 6"
+    orcid: 0000-0001-9249-8624
   - name: Borek Foldyna
     affiliation: 1
+    orcid: 0000-0002-2466-4827
   - given-names: Hugo
     dropping-particle: JWL
     surname: Aerts
     affiliation: "2, 3, 4"
+    orcid: 0000-0002-2122-2003
+
 affiliations:
  - name: Cardiovascular Imaging Research Center, Massachusetts General Hospital, Harvard Medical School, United States of America
    index: 1
@@ -42,7 +46,7 @@ bibliography: paper.bib
 
 # Summary
 
-Lighter is an open-source Python [framework](https://github.com/project-lighter/lighter) for deep learning research that builds upon PyTorch Lightning [@Falcon_PyTorch_Lightning_2019] and the [MONAI Bundle configuration](https://docs.monai.io/en/stable/config_syntax.html#) [@Cardoso_MONAI_An_open-source_2022]. With its declarative YAML-based configuration system that is both transparent and self-documenting, Lighter aims to streamline deep learning research. Researchers define experimental protocols—including neural network architectures, optimization strategies, data pipelines, and evaluation metrics—through structured configuration files, which effectively decouple scientific hypotheses from implementation details. This separation reduces boilerplate code while preserving complete experimental control. Lighter enables reproducibility through comprehensive configuration snapshots that document all experimental parameters and dependencies. A modular adapter system and support for project-specific extensions ensure the extensibility of the framework enabling researchers to implement specialized methodologies without modifying core framework components. By abstracting the engineering complexities of deep learning experimentation, Lighter allows researchers to focus on scientific innovation, accelerate hypothesis testing, and facilitate rigorous validation of research findings across application domains.
+Lighter is an open-source Python [framework](https://github.com/project-lighter/lighter) for deep learning research that builds upon PyTorch Lightning [@Falcon_PyTorch_Lightning_2019] and the [MONAI Bundle configuration](https://docs.monai.io/en/stable/config_syntax.html#) [@Cardoso_MONAI_An_open-source_2022]. With its declarative YAML-based configuration system that is both transparent and self-documenting, Lighter aims to streamline deep learning research. Researchers define experimental protocols—including neural network architectures, optimization strategies, data pipelines, and evaluation metrics—through structured configuration files, effectively decoupling scientific hypotheses from implementation details. This separation reduces boilerplate code while preserving complete experimental control. Lighter enables reproducibility through comprehensive configuration snapshots that document all experimental parameters and dependencies. A modular adapter system and support for project-specific extensions ensure the extensibility of the framework enabling researchers to implement specialized methodologies without modifying core framework components. By abstracting the engineering complexities of deep learning experimentation, Lighter allows researchers to focus on scientific innovation, accelerate hypothesis testing, and facilitate rigorous validation of research findings across application domains.
 
 # Statement of Need
 
@@ -61,19 +65,19 @@ Lighter is designed to address several key challenges in deep learning experimen
 
 # Design
 
-Lighter's architecture is built upon three fundamental components that work together to streamline deep learning experimentation:
+Lighter is built upon three fundamental components:
 
[…]
