+++ paper/paper.bib
@@ -11,7 +11,7 @@
 
 @article{Cardoso_MONAI_An_open-source_2022,
     author = {Cardoso, M. Jorge and Li, Wenqi and Brown, Richard and Ma, Nic and Kerfoot, Eric and Wang, Yiheng and Murray, Benjamin and Myronenko, Andriy and Zhao, Can and Yang, Dong and Nath, Vishwesh and He, Yufan and Xu, Ziyue and Hatamizadeh, Ali and Zhu, Wentao and Liu, Yun and Zheng, Mingxin and Tang, Yucheng and Yang, Isaac and Zephyr, Michael and Hashemian, Behrooz and Alle, Sachidanand and Zalbagi Darestani, Mohammad and Budd, Charlie and Modat, Marc and Vercauteren, Tom and Wang, Guotai and Li, Yiwen and Hu, Yipeng and Fu, Yunguan and Gorman, Benjamin and Johnson, Hans and Genereaux, Brad and Erdal, Barbaros S. and Gupta, Vikash and Diaz-Pinto, Andres and Dourson, Andre and Maier-Hein, Lena and Jaeger, Paul F. and Baumgartner, Michael and Kalpathy-Cramer, Jayashree and Flores, Mona and Kirby, Justin and Cooper, Lee A.D. and Roth, Holger R. and Xu, Daguang and Bericat, David and Floca, Ralf and Zhou, S. Kevin and Shuaib, Haris and Farahani, Keyvan and Maier-Hein, Klaus H. and Aylward, Stephen and Dogra, Prerna and Ourselin, Sebastien and Feng, Andrew},
-    doi = {https://doi.org/10.48550/arXiv.2211.02701},
+    doi = {10.48550/arXiv.2211.02701},
     month = nov,
     title = {{MONAI: An open-source framework for deep learning in healthcare}},
     year = {2022}
@@ -56,7 +56,8 @@
   year      =  2024,
   keywords  = "Biomarkers; Cancer imaging; Tumour biomarkers",
   copyright = "https://creativecommons.org/licenses/by/4.0",
-  language  = "en"
+  language  = "en",
+  doi       = {10.1038/s42256-024-00807-9}
 }
 
 @ARTICLE{Pai2025,
@@ -86,5 +87,6 @@
                   interpretable AI solutions in radiology.",
   year         =  2025,
   primaryClass = "eess.IV",
-  eprint       = "2501.09001"
+  eprint       = "2501.09001",
+  doi          = {10.48550/arXiv.2501.09001}
 }
