### 2025-03-18 ebdbb0b
Shorten the paper
