+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "lighter"
-version = "0.0.3a21"
+version = "0.0.3a22"
 description = "YAML-based automated rapid prototyping framework for deep learning experiments"
 authors = [
     { name = "Ibrahim Hadzic", email = "ibrahimhadzic45@gmail.com" },
