### 2025-03-21 1631fb6
Refactor the paper
