### 2025-04-22 #7884 @fepegar
What are the next steps? Timelines, etc. Do I need to ask the bot to generate my checklist?
