+++ tests/unit/test_engine_resolver.py
@@ -152,47 +152,3 @@ def test_predict_stage_config(resolver):
 
     # Should keep only predict args
     assert list(config.get("args").keys()) == ["predict"]
-
-
-def test_lr_find_stage_config(resolver):
-    """Test configuration for lr_find stage."""
-    config = resolver.get_stage_config(Stage.LR_FIND)
-    system_config = config.get("system")
-
-    # Should keep train and val dataloaders
-    assert "train" in system_config["dataloaders"]
-    assert "val" in system_config["dataloaders"]
-    assert "test" not in system_config["dataloaders"]
-    assert "predict" not in system_config["dataloaders"]
-
-    # Should keep train and val metrics
-    assert "train" in system_config["metrics"]
-    assert "val" in system_config["metrics"]
-    assert "test" not in system_config["metrics"]
-
-    # Should keep optimizer, scheduler, and criterion
-    assert "optimizer" in system_config
-    assert "scheduler" in system_config
-    assert "criterion" in system_config
-
-
-def test_scale_batch_size_stage_config(resolver):
-    """Test configuration for scale_batch_size stage."""
-    config = resolver.get_stage_config(Stage.SCALE_BATCH_SIZE)
-    system_config = config.get("system")
-
-    # Should keep train and val dataloaders
-    assert "train" in system_config["dataloaders"]
-    assert "val" in system_config["dataloaders"]
-    assert "test" not in system_config["dataloaders"]
-    assert "predict" not in system_config["dataloaders"]
-
-    # Should keep train and val metrics
-    assert "train" in system_config["metrics"]
-    assert "val" in system_config["metrics"]
-    assert "test" not in system_config["metrics"]
-
-    # Should keep optimizer, scheduler, and criterion
-    assert "optimizer" in system_config
-    assert "scheduler" in system_config
-    assert "criterion" in system_config
