+++ lighter/utils/types/enums.py
@@ -27,8 +27,6 @@ class Stage(StrEnum):
     VALIDATE = "validate"
     TEST = "test"
     PREDICT = "predict"
-    LR_FIND = "lr_find"
-    SCALE_BATCH_SIZE = "scale_batch_size"
 
 
 class Mode(StrEnum):
