### 2025-04-24 #8101 @editorialbot
[REVIEW]: Lighter: Configuration-Driven Deep Learning
**Submitting author:** @ibro45 (<a href="http://orcid.org/0000-0002-8397-5940">Ibrahim Hadzic</a>)
**Repository:** https://github.com/project-lighter/lighter
**Branch with paper.md** (empty if default branch):
**Version:** v0.0.4
