### 2025-04-27 project-lighter/lighter#152 @ibro45
JOSS Review Tracker
Review: https://github.com/openjournals/joss-reviews/issues/8101
