### 2025-05-08 6904c49
Add missing params arg in tutorial config
