### 2025-05-08 8dedf1b
Fix RuntimeError when computing metrics
