+++ pyproject.toml
@@ -65,6 +65,14 @@ dev = [
     "av==12.0.0",
     "pytest-metadata>=3.1.1",
 ]
+docs = [
+    "mkdocs-autorefs>=1.4.1",
+    "mkdocs-gen-files>=0.5.0",
+    "mkdocs-literate-nav>=0.6.2",
+    "mkdocs-material>=9.6.12",
+    "mkdocs-section-index>=0.3.10",
+    "mkdocstrings[python]>=0.29.1",
+]
 
 [tool.uv]
 # This ensures dev dependencies are always installed by default
