### 2025-05-08 c8b0c7d
Remove redundant config
