+++ pyproject.toml
@@ -11,7 +11,7 @@ authors = [
     { name = "Suraj Pai", email = "b.pai@maastrichtuniversity.nl" },
     { name = "Keno Bressem", email = "kbressem@bwh.harvard.edu" },
 ]
-requires-python = "~=3.10"
+requires-python = ">=3.10"
 readme = "README.md"
 license = "MIT"
 keywords = []
