### 2025-05-11 #8101 @fepegar
@vissarion all we need to do is check (or not) the checkboxes by the deadline?
