### 2025-05-12 #8101 @vissarion
@fepegar Yes, and in case you find an issue (that probably prevents you to check a box) please file one against the submitted repository’s issue tracker.
For more details: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
