### 2025-05-12 d7fdbb0
Improve usage of uv (#159)
* Improve uv usage

* Remove redundant config
