### 2025-05-12 898dab6
Fix Python version in mypy config (#156)
* Fix Python version in mypy config

* Trigger Build
