### 2025-05-12 b5f5908
Remove upper bounds of dependencies (#160)
* Remove upper bounds of dependencies

* Remove upper bound of Python

* Match pyupgrade call

* Apply pyupgrade changes

---------

Co-authored-by: Ibrahim Hadzic <ibrahimhadzic45@gmail.com>
