+++ pyproject.toml
@@ -11,7 +11,7 @@ authors = [
     { name = "Suraj Pai", email = "b.pai@maastrichtuniversity.nl" },
     { name = "Keno Bressem", email = "kbressem@bwh.harvard.edu" },
 ]
-requires-python = "~=3.10"
+requires-python = ">=3.10"
 readme = "README.md"
 license = "MIT"
 keywords = []
@@ -27,19 +27,19 @@ classifiers = [
     "Programming Language :: Python :: 3.12",
 ]
 dependencies = [
-    "torch>=2.1.2,<3",
-    "pandas>=1.5.3,<2",
+    "torch>=2.1.2",
+    "pandas>=1.5.3",
     "numpy<2.0.0",
-    "pytorch-lightning>=2.1.3,<3",
-    "fire>=0.5.0,<0.6",
-    "loguru>=0.6.0,<0.7",
-    "torchmetrics>=1.2.0,<2",
-    "tensorboard>=2.11.2,<3",
-    "requests>=2.31.0,<3",
+    "pytorch-lightning>=2.1.3",
+    "fire>=0.5.0",
+    "loguru>=0.6.0",
+    "torchmetrics>=1.2.0",
+    "tensorboard>=2.11.2",
+    "requests>=2.31.0",
     "monai>=1.4.0",
-    "rich>=13.7.0,<14",
-    "torchvision>=0.20.0,<0.21",
-    "lightly>=1.5.13,<2",
+    "rich>=13.7.0",
+    "torchvision>=0.20.0",
+    "lightly>=1.5.13",
     "cerberus>=1.3.7",
 ]
 
@@ -53,13 +53,13 @@ lighter = "lighter.engine.runner:cli"
 [dependency-groups]
 # Dependencies for dev/CI tasks
 dev = [
-    "typing-extensions>=4.4.0,<5",
-    "aiohttp>=3.8.3,<4",
-    "pytest>=7.4.0,<8",
-    "pytest-html>=3.2.0,<4",
-    "pytest-cov>=4.0.0,<5",
-    "coverage>=7.0.5,<8",
-    "coverage-badge>=1.1.0,<2",
+    "typing-extensions>=4.4.0",
+    "aiohttp>=3.8.3",
+    "pytest>=7.4.0",
+    "pytest-html>=3.2.0",
+    "pytest-cov>=4.0.0",
+    "coverage>=7.0.5",
+    "coverage-badge>=1.1.0",
     "pylint>=3.3.3",
     "itk>=5.4.0",
     "av==12.0.0",
