### 2025-05-12 10a3b51
Remove lightly (#155)
* Remove lightly from dependencies

* Trigger Build

---------

Co-authored-by: Ibrahim Hadzic <ibrahimhadzic45@gmail.com>
