+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,3 +1,4 @@
 # Configuration: https://help.github.com/en/github/building-a-strong-community/configuring-issue-templates-for-your-repository
 
-blank_issues_enabled: false
+blank_issues_enabled: true
+
