### 2025-05-13 ec0960d
Fix metric registration
