+++ lighter/utils/types/containers.py
@@ -42,7 +42,7 @@ class Metrics:
         self.test = self._convert_to_collection(self.test)
 
     def _convert_to_collection(self, x):
-        if  x is not None and not isinstance(x, MetricCollection):
+        if x is not None and not isinstance(x, MetricCollection):
             return MetricCollection(x)
         return x
 
