+++ .pre-commit-config.yaml
@@ -6,6 +6,7 @@ repos:
     hooks:
       - id: check-yaml
       - id: end-of-file-fixer
+      - id: trailing-whitespace
         exclude: LICENSE
 
   - repo: local
