### 2025-05-14 bb19823
Fix classification tutorial (#157)
* Add missing params arg in tutorial config

* Fix location of config.yaml

* Fix project path

* Improve understanding of the project field

* Add the missing line in the dir tree

* Update the link to How-To guides

---------

Co-authored-by: Ibrahim Hadzic <ibrahimhadzic45@gmail.com>
