### 2025-05-14 add9651
Fix metric registration (#162)
* Fix metric registration

* Fix style

* Fix test
