### 2025-05-14 f13ce09
Updating pre-commit hook (#164)
* Updating pre-commit hook

* Update .pre-commit-config.yaml

---------

Co-authored-by: Ibrahim Hadzic <ibrahimhadzic45@gmail.com>
