+++ .pre-commit-config.yaml
@@ -7,6 +7,7 @@ repos:
       - id: check-yaml
       - id: end-of-file-fixer
         exclude: LICENSE
+      - id: trailing-whitespace
 
   - repo: local
     hooks:
