### 2025-05-28 9c1f403
Redo intro, improve monai config description. Shorten Statement-of-Need. Elaborate Adapters.
