### 2025-06-02 6c7e01f
Add state of the field. Improve adapters and project module sections.
