+++ paper/paper.bib
@@ -90,3 +90,45 @@
   eprint       = "2501.09001",
   doi          = {10.48550/arXiv.2501.09001}
 }
+
+@article{Ludwig,
+  author       = {Piero Molino and
+                  Yaroslav Dudin and
+                  Sai Sumanth Miryala},
+  title        = {Ludwig: a type-based declarative deep learning toolbox},
+  journal      = {CoRR},
+  volume       = {abs/1909.07930},
+  year         = {2019},
+  url          = {http://arxiv.org/abs/1909.07930},
+  eprinttype    = {arXiv},
+  eprint       = {1909.07930},
+  timestamp    = {Tue, 24 Sep 2019 11:33:51 +0200},
+  biburl       = {https://dblp.org/rec/journals/corr/abs-1909-07930.bib},
+  bibsource    = {dblp computer science bibliography, https://dblp.org}
+}
+
+@software{Quadra,
+	title = {orobix/quadra},
+	url = {https://github.com/orobix/quadra},
+	author = {Mammana, Lorenzo and Malli, Refik Can and Polidori, Alessandro and ebernatene},
+	date = {2025-05-20},
+	year = {2025},
+	month = {5},
+	day = {20},
+}
+
+@article{Gandlf,
+    author={Pati, Sarthak and Thakur, Siddhesh P. and Hamamc{\i}, {\.{I}}brahim Ethem and Baid, Ujjwal and Baheti, Bhakti and Bhalerao, Megh and G{\"u}ley, Orhun and Mouchtaris, Sofia and Lang, David and Thermos, Spyridon and Gotkowski, Karol and Gonz{\'a}lez, Camila and Grenko, Caleb and Getka, Alexander and Edwards, Brandon and Sheller, Micah and Wu, Junwen and Karkada, Deepthi and Panchumarthy, Ravi and Ahluwalia, Vinayak and Zou, Chunrui and Bashyam, Vishnu and Li, Yuemeng and Haghighi, Babak and Chitalia, Rhea and Abousamra, Shahira and Kurc, Tahsin M. and Gastounioti, Aimilia and Er, Sezgin and Bergman, Mark and Saltz, Joel H. and Fan, Yong and Shah, Prashant and Mukhopadhyay, Anirban and Tsaftaris, Sotirios A. and Menze, Bjoern and Davatzikos, Christos and Kontos, Despina and Karargyris, Alexandros and Umeton, Renato and Mattson, Peter and Bakas, Spyridon},
+    title={GaNDLF: the generally nuanced deep learning framework for scalable end-to-end clinical workflows},
+    journal={Communications Engineering},
+    year={2023},
+    month={May},
+    day={16},
+    volume={2},
+    number={1},
+    pages={23},
+    abstract={Deep Learning (DL) has the potential to optimize machine learning in both the scientific and clinical communities. However, greater expertise is required to develop DL algorithms, and the variability of implementations hinders their reproducibility, translation, and deployment. Here we present the community-driven Generally Nuanced Deep Learning Framework (GaNDLF), with the goal of lowering these barriers. GaNDLF makes the mechanism of DL development, training, and inference more stable, reproducible, interpretable, and scalable, without requiring an extensive technical background. GaNDLF aims to provide an end-to-end solution for all DL-related tasks in computational precision medicine. We demonstrate the ability of GaNDLF to analyze both radiology and histology images, with built-in support for k-fold cross-validation, data augmentation, multiple modalities and output classes. Our quantitative performance evaluation on numerous use cases, anatomies, and computational tasks supports GaNDLF as a robust application framework for deployment in clinical workflows.},
+    issn={2731-3395},
+    doi={10.1038/s44172-023-00066-3},
+    url={https://doi.org/10.1038/s44172-023-00066-3}
+}
