### 2025-06-04 135cc00
Refine "State of the Field" section for clarity and precision. Update descriptions of abstraction levels and pipeline flexibility in Lighter.
