### 2025-06-04 a14d8a3
Update paper as per reviewer's comments (#166)
* Redo intro, improve monai config description. Shorten Statement-of-Need. Elaborate Adapters.

* Add state of the field. Improve adapters and project module sections.

* Refine "State of the Field" section for clarity and precision. Update descriptions of abstraction levels and pipeline flexibility in Lighter.

* Update paper.pdf
