+++ src/lighter/__init__.py
@@ -8,6 +8,8 @@ from .utils.logging import _setup_logging
 
 _setup_logging()
 
-from .engine.config import Config
-from .engine.runner import Runner
-from .system import System
+from .engine.config import Config  # noqa: E402
+from .engine.runner import Runner  # noqa: E402
+from .system import System  # noqa: E402
+
+__all__ = ["Config", "Runner", "System"]
