+++ src/lighter/callbacks/__init__.py
@@ -1,3 +1,5 @@
 from .freezer import Freezer
 from .writer.file import FileWriter
 from .writer.table import TableWriter
+
+__all__ = ["Freezer", "FileWriter", "TableWriter"]
