### 2025-07-09 5ca42df
Adopt `src/` layout and revamp CI/CD
Migrate all source code from `lighter/` to `src/lighter/` to align with modern Python packaging standards.
Overhaul GitHub Actions by replacing `ci.yml` with dedicated workflows for:
- `tests.yml`: Run unit and integration tests.
- `code_quality.yml`: Enforce linting and formatting.
- `publish.yml`: Handle package publishing.
- `check_pull_request_title.yml`: Ensure PR title compliance.
Transition from `Makefile` to `justfile` for task automation.
Add `tox.ini` for multi-environment testing.
Update `pyproject.toml`, `.pre-commit-config.yaml`, and documentat
[…]
