### 2025-07-09 #8101 @ibro45
Thank you, @vissarion!

Version: v0.0.4 https://github.com/project-lighter/lighter/releases/tag/v0.0.4
DOI (Zenodo): https://doi.org/10.5281/zenodo.15844257
